feat: Phase 7 Monitoring & Alerting System
- Implemented K8s metrics collector using k8s.io/metrics
- Implemented alerting rule engine with threshold and duration support
- Implemented notifications via Webhook and Email
- Implemented SQLite storage for metrics, alerts, and silences
- Added Monitoring API and CLI commands
- Refactored plugin management to support legacy static plugins alongside dynamic ones
- Implemented LegacyPluginAdapter to bridge legacy plugins for metrics collection
- Fixed web crawler config and implementation issues
- Enabled middleware collector registration in server.go

+++ internal/knowledge/crawler/web_crawler_impl.go
@@ -168,7 +168,7 @@ func (c *advancedCrawler) CrawlWithFilter(ctx context.Context, startURL string) 
 		title := doc.Find("title").Text()
 
 		score := c.scorer.Score(cleanedContent)
-		if score >= c.config.Quality.MinScore {
+		if float64(score) >= c.config.Quality.MinScore {
 			doc := &CrawledDocument{
 				URL:     r.Request.URL.String(),
 				Content: cleanedContent,
